pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "autotrade"
version = "0.1.0"
description = "KOSPI/KOSDAQ 장중 단타 자동매매 시스템"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.9",
    "websockets>=12.0",
    "pandas>=2.0",
    "pykrx>=1.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "pyyaml>=6.0",
    "loguru>=0.7",
    "fastapi>=0.110",
    "uvicorn>=0.27",
    "python-telegram-bot>=20.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.23",
]

[project.scripts]
autotrade = "run:main"

[tool.setuptools.packages.find]
include = ["src*", "config*"]
